requirements.txt
# Fundamental Math & Tensor
numpy==1.26.4
scipy==1.11.4

# Audio Analysis & Diarization
pyannote.audio==3.1.1
librosa==0.10.1
soundfile==0.12.1

# Speech-to-Text Engine
faster-whisper==1.2.1

# Content Acquisition
yt-dlp==2024.03.10

# Utilities
python-dotenv==1.0.1
pydantic==2.6.4
httpx==0.27.0
fastapi==0.110.0
uvicorn==0.28.0
tqdm==4.66.2
